requirements.txt
# Discord API
aiohttp
uvloop>=0.16; os_name != 'nt'
hikari
# hikari[speedups]
# git+https://github.com/FasterSpeeding/Tanjun.git
hikari-tanjun
hikari-yuyo
# git+https://github.com/thesadru/tanchi.git
# hikari-tanchi

# Data management
aiofiles
python-dotenv
asyncpg
hikari-sake

# Voice
lavasnek_rs; os_name != 'nt'

# extras
pygount
distro
psutil
